Former-commit-id: 4fd0e80b819837219d277bfa59b24dfac2804a59

+++ Arduino/startROSArduino.sh
@@ -2,6 +2,12 @@
 
 #FIRST AND ONLY ARGUMENT IS THE ACM port number
 
+if [ $# -eq 0 ]
+    then
+        echo "No arguments provided"
+        exit 1
+fi
+
 echo "***Kill roscore if already running"
 killall roscore
 
@@ -14,5 +20,5 @@ sleep 5
 echo "***Starting rosserial and binding to /dev/ttyACM(PORT)"
 xterm -e rosrun rosserial_python serial_node.py /dev/ttyACM$1 &
 
-echo "***Publish arduino feedback"
+echo "***Publish arduino feedback" &
 #xterm -e rostopic echo /arduino_feeback &
